requirements.txt
aniso8601==10.0.0
attrs==24.3.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
Flask==3.1.0
Flask-JWT-Extended==4.7.1
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
importlib_resources==6.5.2
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
pycparser==2.22
PyJWT==2.10.1
pytz==2024.2
referencing==0.36.1
rpds-py==0.22.3
SQLAlchemy==2.0.37
typing_extensions==4.12.2
Werkzeug==3.1.3
python-dotenv==1.0.1